Prevent developer from saving DEV edition from VBE
Added a private event handler for the toolkit's BeforeSave event.  This
handler checks a public boolean flag in the bootstrap module to see if
the current toolkit edition can be saved or not.  Saving is disabled by
default, and is only enabled when building the production edition, or
for the special NO-LOAD file name.  Fixes GH #5.